refactor: clean up useVaults hooks
